+++ lib/hubspot/contact_properties.rb
@@ -1,4 +1,4 @@
-module Hubspot
+module DeprecatedHubspot
   class ContactProperties < Properties
 
     ALL_PROPERTIES_PATH  = '/contacts/v2/properties'