+++ lib/hubspot/deprecator.rb
@@ -1,4 +1,4 @@
-module Hubspot
+module DeprecatedHubspot
   class Deprecator
     def self.build(version: "1.0")
       ActiveSupport::Deprecation.new(version, "hubspot-ruby")